+++ Cedai_Engine/src/model/Vertex.hpp
@@ -11,8 +11,8 @@ namespace cd {
 		glm::ivec4 boneIndices = { -1, -1, -1, -1 };
 		glm::vec4 boneWeights = { 0, 0, 0, 0 };
 
-		static std::array<int, 3> getOffsets() {
-			return std::array<int, 3> {
+		static std::array<size_t, 3> getOffsets() {
+			return std::array<size_t, 3> {
 					offsetof(Vertex, position),
 					offsetof(Vertex, boneIndices),
 					offsetof(Vertex, boneWeights)
